dlx: remove init.qcom.sh
First of all, the service definition for "startup" is wrong.  It
is constantly being restarted by init burning CPU.

Second, if you manually trace through all the conditions, it
isn't doing anything except chown/chmod'ing a file that doesn't
exists and disabling some default routes for IPv6 because
user-space will take care of it.  I doubt that this is referring
to our user-space...

Change-Id: I0dbf1b18a3b35713df04c71209515a8c0ffad693

Conflicts:

	full_dlx.mk